Fixes S3 predict function
Changes ~ 2 years ago to holstius/CALINE3 mean that CALINE3_RECEPTOR_TOTALS does not exist anymore; only CAL3RXM and CAL3RXL are possible targets for the do.call() via the CALINE3 package. Thus, running the San Francisco vignette dies on the do.call() in the predict() line. This adds a parameter to the predict function, defaulting to CAL3RXL, which allows the prediction to function, while still allowing the user to select between RXL and RXM models.